Правки и улучшения - Теперь параметры аудио потока не передаются из nodejs а существуют напрямую в Rust, фильтры и ресурс все еще передаются из nodejs - Мелкие правки Scheduler, перенос параметров в constants.rs - Обновление @napi-rs/cli до 3.10.4, @types/node до 26.6.2 - Разделение планировщика тиков     scheduler     ├── mod.rs          # Публичный интерфейс     ├── timer.rs        # Прецизионный таймер с адаптивным спином     ├── registry.rs     # Потокобезопасное хранилище сессий     └── telemetry.rs    # Метрики и EMA - Разделение AudioEngine на модули     mod.rs          — структура, поля, конструктор, счётчик живых     ├── constants.rs    # все константы     ├── reader.rs       # тело потока чтения (чистая функция)     ├── start.rs        # start()     ├── lifecycle.rs    # destroy(), force_destroy(), Drop     └── buffer_ops.rs   # get_size / get_position / get_packets / add_packets / can_accept*

+++ rs/src/structures/crypto/aes.rs
@@ -1,4 +1,4 @@
-use crate::structures::timers::scheduler::TICK_INTERVAL_MS;
+use crate::structures::timers::scheduler::constants::TICK_INTERVAL_MS;
 use std::{
     sync::{
         atomic::{AtomicU16, AtomicU32, Ordering},
@@ -65,7 +65,7 @@ const RTP_PAYLOAD_TYPE_OPUS: u8 = 120;
 ///
 /// Для Opus с частотой дискретизации 48 кГц и кадрами по 20 мс получаем 960 семплов.
 /// `TICK_INTERVAL_MS` — интервал цикла отправки (20 мс).
-const TIMESTAMP_INC: u32 = 48000 * TICK_INTERVAL_MS / 1000;
+const TIMESTAMP_INC: u32 = 48000 * (TICK_INTERVAL_MS as u32) / 1000;
 
 // ============================================================================
 // Ошибки криптографических операций
